+++ elements/linuxmodule/rtcycles.cc
@@ -35,7 +35,7 @@ RTCycles::~RTCycles()
 void
 RTCycles::push(int, Packet *p)
 {
-  unsigned long long c = click_get_cycles();
+  uint64_t c = click_get_cycles();
   output(0).push(p);
   _accum += click_get_cycles() - c;
   _npackets++;
@@ -44,7 +44,7 @@ RTCycles::push(int, Packet *p)
 Packet *
 RTCycles::pull(int)
 {
-  unsigned long long c = click_get_cycles();
+  uint64_t c = click_get_cycles();
   Packet *p = input(0).pull();
   _accum += click_get_cycles() - c;
   if (p) _npackets++;
